% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationSPoTApi.r
\name{PreparationSPoTApi_get_preparation_versions}
\alias{PreparationSPoTApi_get_preparation_versions}
\title{Retrieve a list of object versions by ID}
\usage{
PreparationSPoTApi_get_preparation_versions(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve a list of object versions by ID
}
