% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationSPoTApi.r
\name{PreparationSPoTApi_search_preparations}
\alias{PreparationSPoTApi_search_preparations}
\title{List or search for preparation metadata objects}
\usage{
PreparationSPoTApi_search_preparations(
  filter,
  query,
  search_specific_terms,
  returned_metadata_fields,
  sort,
  page_limit,
  page_offset,
  ...
)
}
\arguments{
\item{filter}{Filter by preparation metadata (key-value metadata pair(s)). E.g. \code{Digestion=Trypsin}}

\item{query}{Search for preparation objects via a full-text query over all preparation metadata fields. E.g. \code{"reversed-phase liquid chromatography"}}

\item{search_specific_terms}{If the full-text query term is present in an ODM dictionary, enabling this parameter will modify the query to include child terms of the full-text query.  For example, the search query "Body fluid" can be expanded to include the term "Blood" (a child term of  "Body fluid") so files containing either "Body fluid" or  "Blood"  in their metadata will be returned in the search results.  The parent-child relationship is defined by the key "broaders" in the dictionary.  If the full query term is not present in a dictionary then this parameter has no effect.}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}

\item{sort}{Attribute to sort by, with optional ascending/descending marker, of the form \code{"[+|-]<key_name>"}.  No marker or \code{"+"} indicates ascending sort, and \code{"-"} indicates descending sort.  *Default:* sort by ID in ascending order.}

\item{page_limit}{Maximum number of results to return per page (see Paging above). This value must be between 0 and 2000 (inclusive). The default is 2000.}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}
}
\description{
List or search for preparation metadata objects
}
